italy-tourism-insights/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
statsmodels==0.14.0
prophet==1.1.5
scipy==1.11.4
plotly==5.18.0
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
groq==0.4.2
